requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.10
pillow==11.3.0
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
